diploma version

+++ Engine/Source/VeiM/Core/Source/Windows/WindowsUtils.cpp
@@ -10,7 +10,7 @@
 
 namespace VeiM
 {
-	float Time::GetTime()
+	double Time::GetTime()
 	{
 		return glfwGetTime();
 	}
